+++ Imaging/vtkImageRGBToHSI.h
@@ -34,14 +34,14 @@ public:
   // Hue is an angle. Maximum specifies when it maps back to 0.  HueMaximum
   // defaults to 255 instead of 2PI, because unsigned char is expected as
   // input.  Maximum also specifies the maximum of the Saturation.
-  vtkSetMacro(Maximum,float);
-  vtkGetMacro(Maximum,float);
+  vtkSetMacro(Maximum,double);
+  vtkGetMacro(Maximum,double);
   
 protected:
   vtkImageRGBToHSI();
   ~vtkImageRGBToHSI() {};
 
-  float Maximum;
+  double Maximum;
   
   void ThreadedExecute(vtkImageData *inData, vtkImageData *outData,
                        int ext[6], int id);
